// TrashMob/client-app/src/components/Models/SponsoredAdoptionComplianceStats.tsx
export interface SponsoredAdoptionComplianceStats {
    totalSponsoredAdoptions: number;
    activeAdoptions: number;
    expiredAdoptions: number;
    terminatedAdoptions: number;
    adoptionsOnSchedule: number;
    adoptionsOverdue: number;
    totalCleanupLogs: number;
    totalWeightCollectedPounds: number;
    totalBagsCollected: number;
    complianceRate: number;
}